NoneType condition for run()
run() actually returns a NoneType object if no documents match the query. Should this be explicit in the API documentation since JSON or a cursor aren't the only return values for this function? 

Also, to be super nitpicky, when I'm running queries on Python 2.7 it's actually returning a Python 'dict' rather than a JSON object (I like this behavior, but it might be good to document it here rather than saying it's JSON). 

What do you think?